Made install script check if the files exitst before installing them

+++ install.sh
@@ -4,12 +4,20 @@ if [ $EUID != 0 ]; then
 	exit $?
 fi
 
-ln -s ~/.dotfiles/tmux.conf ~/.tmux.conf
-ln -s ~/.dotfiles/zshrc ~/.zshrc
-ln -s ~/.dotfiles/oh-my-zsh ~/.oh-my-zsh
-ln -s ~/.dotfiles/vimrc ~/.vimrc
-ln -s ~/.dotfiles/vim ~/.vim
-ln -s ~/.dotfiles/gitignore ~/.gitignore
+if [ ! -f ~/.tmux.conf ]; then ln -s ~/.dotfiles/tmux.conf ~/.tmux.conf; fi
+if [ ! -f ~/.zshrc ]; then ln -s ~/.dotfiles/zshrc ~/.zshrc; fi
+if [ ! -f ~/.vimrc ]; then ln -s ~/.dotfiles/vimrc ~/.vimrc; fi
+if [ ! -d ~/.vim ]; then ln -s ~/.dotfiles/vim ~/.vim; fi
+if [ ! -f ~/.gitignore ]; then ln -s ~/.dotfiles/gitignore ~/.gitignore; fi
 
-curl -fsSL https://raw.github.com/mislav/dotfiles/1500cd2/bin/tmux-vim-select-pane -o /usr/local/bin/tmux-vim-select-pane
-chmod +x /usr/local/bin/tmux-vim-select-pane
+#ln -s ~/.dotfiles/zshrc ~/.zshrc
+#ln -s ~/.dotfiles/vimrc ~/.vimrc
+#ln -s ~/.dotfiles/vim ~/.vim
+#ln -s ~/.dotfiles/gitignore ~/.gitignore
+
+if [ ! -f /usr/local/bin/tmux-vim-select-pane ]; then
+	curl -fsSL https://raw.github.com/mislav/dotfiles/1500cd2/bin/tmux-vim-select-pane -o /usr/local/bin/tmux-vim-select-pane;
+	chmod +x /usr/local/bin/tmux-vim-select-pane;
+fi
+
+echo "Done installing..."
